実行部分とライブラリ部分を分離
