+++ src/lib.rs
@@ -1,0 +1,7 @@
+/// 標準入力の簡素化モジュール
+pub mod input; 
+pub use input::*;
+/// 任意の素数を法とする変数
+pub mod modint;
+pub use modint::*;
+
